Fix EnumField deconstruction and cloning.
The default value was not preserved when the field was cloned and the
instance was created without the enum argument.

Additionally, the choices argument was not present in values from
deconstruct() when the instance was created without the enum argument.

Both these problems caused issues during generating migrations: default
values was not present because of the issue with default and fields were
always marked as changed even when there were no changes.