Fixed params instance re-use in DBType.
When no params were given when creating DBType instance, no the params
value was not assigned to self and static params were used. However,
this means that the instance of empty array was shared between
all instances and this caused problems during migration execution if
the migration contained multiple Enums with a default value: additional
parameter for default value was added to the shared array.